==> project/code/2.4_algorithm_lasso.ipynb <==
# Lasso Algorithm

Lasso is a regression model that uses shrinkage to bring data points towards the center, similar to the mean value of all the data points. Lasso stands for Least Absolute Shrinkage and Selection Operator. It is a regularized linear model with penalty term lambda to minimize the error. Parameter penalization controls overfitting the input data by shrinking variable coefficients to 0. Essentially this makes the variables no effect in the model, hence reduces the dimensions.

import os
import numpy as np 
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import Lasso
import matplotlib.pyplot as plt
import pandas as pd
from sklearn.metrics import mean_squared_error

print(os.getcwd())
print("")

train = pd.read_csv("../data/train_after_feature_engineering.csv")
test = pd.read_csv("../data/test_after_feature_engineering.csv")


print ('The train data has {0} rows and {1} columns'.format(train.shape[0],train.shape[1]))    
print ('The test data has {0} rows and {1} columns'.format(test.shape[0],test.shape[1]))


np.random.seed(1234)

#found this best alpha value through cross-validation
_best_alpha = 0.0001

_lasso_algo = Lasso(alpha = _best_alpha, max_iter = 50000)

# Fit the Model

target_vector = pd.DataFrame(index = train.index, columns=["SalePrice"])
target_vector["SalePrice"] = train["SalePrice"]

target_vector["SalePrice"] = np.log1p(target_vector["SalePrice"]) # log(SalePrice) + 1

train.drop(['SalePrice'], axis=1, inplace=True)

_lasso_algo.fit(train, target_vector)


# Predict the Test Sale Price

y_train = target_vector
y_train_pred = _lasso_algo.predict(train)
    
rmse_train = np.sqrt(mean_squared_error(y_train,y_train_pred))

print("Lasso score on training set: ", rmse_train)

y_test_pred = _lasso_algo.predict(test)

print(y_test_pred[5:])

# Save Predictions

df_predict = pd.DataFrame({'Id': test["Id"], 'SalePrice': np.exp(y_test_pred) - 1.0})
#df_predict = pd.DataFrame({'Id': id_vector, 'SalePrice': sale_price_vector})

df_predict.to_csv('../data/kaggle_python_lasso.csv', header=True, index=False)

print("..file saved")
