==> project/code/3_ensemble_kaggle_submission.ipynb <==
# Ensemble and Kaggle Submission

We can create a robust predictive model with better accuracy by merging two or more machine learning algorithms. This technique is called model ensembling. Ensembled algorithms may be similar in functionality or may entirely be different from each other. Individual algorithms may not perform great but by ensembling them, the overall system can offer much better performance and accuracy. Variations in the predicting logic in each of these individual algorithms will bring unbiasedness into the unified model. Bagging, boosting and stacking are popular ensembling techniques. Many of the advanced machine learning algorithms use ensembled approaches to achieve accurate classifications or predictions. Random Forest uses bagging, XGBoost uses boosting and Neural Network applies stacking ensembling techniques. For the kaggle submission, we have created an ensembled model by averaging Sale Price of the top 3 performing ensembled algorithms - XGBoost, Lasso and Neural Network.  Following are the kaggle public scores that we have achieved by submitting all the six algorithms along with the ensembled model. As predicted, ensembled model has scored better compared to the individual algorithms. By applying advanced machine learning algorithms, we have placed our scores within top 15% of the competition. Following table displays each algorithm and the root-mean-square error (RMSE) along with the kaggle score.
	

import os
import pandas as pd
import numpy as np 

print(os.getcwd())
print("")

xgb_prediction = pd.read_csv("../data/kaggle_python_xgboost.csv")
random_forest_prediction = pd.read_csv("../data/kaggle_python_random_forest.csv")
lasso_prediction = pd.read_csv("../data/kaggle_python_lasso.csv")


#ensemble best 3 algorithms
ensemble_prediction = (xgb_prediction["SalePrice"] + 
                       random_forest_prediction["SalePrice"] + 
                       lasso_prediction["SalePrice"]) / 3

df_predict = pd.DataFrame({'Id': xgb_prediction["Id"], 
                           'SalePrice': ensemble_prediction})

print(df_predict.head())

df_predict.to_csv('../data/kaggle_python_ensemble.csv', header=True, index=False)

print("...file saved")
